python_requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
pandas==2.1.3
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
httpx==0.25.2
requests==2.31.0
requests-toolbelt==1.0.0
python-dotenv==1.0.0